correcion pdf repetidos

+++ controller/administrativo/Modificar_Poliza_Individual.php
@@ -27,7 +27,7 @@ if(isset($_POST["id_info"]) && isset($_POST["servicio_tipo"])){
     $polizas_in = json_decode($_POST["polizas_in"]);
     $path = "../comprobantes/administrativo/polizas/";
 
-    
+
   
     $resultado_procesado = $objeto->modificar_polizas($_POST["id"],$polizas_in);
 
@@ -40,49 +40,56 @@ if(isset($_POST["id_info"]) && isset($_POST["servicio_tipo"])){
 
     }else{
        
-        $index_files = 0;
-        for ($i=0; $i < count($resultado_procesado) ; $i++) { 
+        if(isset($_FILES['archivos'])){
 
-            $poliza = $resultado_procesado[$i];
-            
-            
-            if($poliza[0] == "new"){
-                $file = $path.$poliza[1].".pdf";
-                $tmpName = $_FILES['archivos']['tmp_name'][$index_files];
-                move_uploaded_file($tmpName, $file);
+            $index_files = 0;
+            for ($i=0; $i < count($resultado_procesado) ; $i++) { 
 
-                $index_files++;
-    
-            }else if($poliza[0] == "update"){
-                $hayArchivo = $poliza[6];
-                $file = $path.$poliza[1].".pdf";
-
-                if($hayArchivo == "si"){
-                    if(file_exists($file)){
-                    
-                        unlink($file);
-                        $tmpName = $_FILES['archivos']['tmp_name'][$index_files];
-                        move_uploaded_file($tmpName, $file);
-                    }else{
-                        $tmpName = $_FILES['archivos']['tmp_name'][$index_files];
-                        move_uploaded_file($tmpName, $file);
-                    }
-                    
-                }
-                $index_files++;
+                $poliza = $resultado_procesado[$i];
                 
-    
-            }else{
-                $file = $path.$poliza[1].".pdf";
-                if(file_exists($file)){
-                    unlink($file);
+                
+                if($poliza[0] == "new"){
+                    $file = $path.$poliza[1].".pdf";
+                    $tmpName = $_FILES['archivos']['tmp_name'][$index_files];
+                    move_uploaded_file($tmpName, $file);
+
+                    $index_files++;
+        
+                }else if($poliza[0] == "update"){
+                    $hayArchivo = $poliza[6];
+                    $file = $path.$poliza[1].".pdf";
+
+                    if($hayArchivo == "si"){
+                        if(file_exists($file)){
+                        
+                            unlink($file);
+                            $tmpName = $_FILES['archivos']['tmp_name'][$index_files];
+                            move_uploaded_file($tmpName, $file);
+                        }else{
+                            $tmpName = $_FILES['archivos']['tmp_name'][$index_files];
+                            move_uploaded_file($tmpName, $file);
+                        }
+
+                        $index_files++;
+                        
+                    }
+                   
+                    
+        
+                }else{
+                    $file = $path.$poliza[1].".pdf";
+                    if(file_exists($file)){
+                        unlink($file);
+                    }
                 }
+                
             }
-            
+
+           
+
         }
-
+        
         $datos = true;
-
     }
